requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.8.0+cu111
torchvision==0.9.0+cu111
torchaudio===0.8.0 
scikit-video
matplotlib
pyyaml
argparse
numpy
h5py
opencv-python
imageio
more-itertools